Finish blacklist

+++ reddit.py
@@ -41,12 +41,12 @@ def getMessages():
 
 
 def markMessageRead(message):
-	if whitelist is None or message.author in whitelist:
+	if whitelist is None:
 		message.mark_as_read()
 
 
 def replyMessage(message, body):
-	if whitelist is not None and message.author not in whitelist:
+	if whitelist is not None and str(message.author) not in whitelist:
 		return True
 	try:
 		message.reply(body)
@@ -65,9 +65,16 @@ def deleteComment(id=None, comment=None):
 		return True
 	try:
 		if id is not None:
-			reddit.get_info(thing_id='t1_' + id).delete()
+			idComment = reddit.get_info(thing_id='t1_' + id)
+			if str(idComment.author).lower() == globals.ACCOUNT_NAME.lower():
+				idComment.delete()
+			else:
+				log.warning("Skipping comment delete, not author")
 		if comment is not None:
-			comment.delete()
+			if str(comment.author).lower() == globals.ACCOUNT_NAME.lower():
+				comment.delete()
+			else:
+				log.warning("Skipping comment delete, not author")
 		return True
 	except Exception as err:
 		log.warning(traceback.format_exc())
@@ -77,7 +84,7 @@ def deleteComment(id=None, comment=None):
 def replyComment(id, message):
 	try:
 		parent = reddit.get_info(thing_id='t1_' + id)
-		if whitelist is not None and parent.author not in whitelist:
+		if whitelist is not None and str(parent.author) not in whitelist:
 			return None
 		resultComment = parent.reply(message)
 		return resultComment.id
@@ -90,8 +97,13 @@ def editComment(id, message):
 	if whitelist is not None and not len(whitelist):
 		return True
 	try:
-		reddit.get_info(thing_id='t1_' + id).edit(message)
-		return True
+		comment = reddit.get_info(thing_id='t1_' + id)
+		if str(comment.author).lower() == globals.ACCOUNT_NAME.lower():
+			comment.edit(message)
+			return True
+		else:
+			log.warning("Skipping comment edit, not author")
+			return True
 	except Exception as err:
 		log.warning(traceback.format_exc())
 		return False
